+++ tests/dummy_unit_test.py
@@ -4,6 +4,7 @@ import logging
 from pathlib import Path
 
 import pytest
+
 from cheshire.dummy import do_something
 
 logger = logging.getLogger(__name__)
@@ -34,7 +35,6 @@ def test_something(a: int, b: int, expected_c: int) -> None:
     assert c == expected_c  # nosec: B101
 
 
-@pytest.mark.usefixtures("pudl_access_key_setup")
 def test_use_a_table_from_pudl() -> None:
     """Test that we can run a function that uses a table from the PUDL GCS bucket."""
     from cheshire.dummy_pudl import use_a_table_from_pudl